move functions within classes
